feat:新增设备门禁监控

+++ src/main/java/com/networkinto/facility/common/dto/HumanFaceDto.java
@@ -73,5 +73,5 @@ public class HumanFaceDto {
     /**
      * 下发失败次数
      */
-    private Long noNum;
+    private int noNum;
 }
